feat(translate): 一键完成看词选义 reflact(word): 重构word.js，拆分功能
