+++ README.md
@@ -2,9 +2,10 @@
 ## 功能
 - [x] 一键查看听说作业答案
 - [x] 一键查看单词选择题答案
-- [x] 单词作业自动下一个
+- [x] 单词朗读自动下一个
 - [x] 模仿朗读开启声音
 - [x] 一键修改分数
+- [x] 一键完成看词选义
 - [x] VIP破解(测试中)
 - [ ] ………
