+++ package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "element-plus": "^2.11.1",
+    "element-plus": "2.11.3",
     "vue": "^3.5.20"
   },
   "devDependencies": {
@@ -18,7 +18,7 @@
     "eslint-plugin-vue": "^10.4.0",
     "unplugin-auto-import": "^20.0.0",
     "unplugin-vue-components": "^29.0.0",
-    "vite": "^7.1.3",
+    "vite": "7.1.6",
     "vite-plugin-monkey": "^7.1.2"
   }
 }
